[sparse-fsm-encode] Switch to Safe Rust Encoding
In Rust, unlike C, enum values which are not a named variant's value
cause undefined behaviour. This means they cannot be treated like an
integer, like they might be in C.

We use these sparse encodings with the understanding that the code may
see any possible encoding value, but only certain values are defined to
have specific behaviour (and all other values cause errors).

We need to not trigger undefined behaviour in our code, so we choose a
more manual type representation for these Rust enums, which uses the
`newtype` struct, and named constants.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>