[rom_ext] Document Usage Constraint Value and Signing Scheme
This is the devised scheme based on the scheme outlined in the
Software/Security Sync on 2020-10-07, and developed further in meetings
on 2020-11-02 and 2020-11-05.

This change introduces a method for calculating a "device usage value"
based on device state and the rom_ext's "usage constraint" value. That
is used in the signing scheme as one of the inputs for signature
generation. These values both now have defined sizes.

This change also introduces a "system state value" which can be used to
validate the system has come up in the intended configuration to boot a
specific rom_ext.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>