[otbn] Wrap OTBN model with definition
This commit is related to #4144

otbn_model is in the dependency list in otbn.core. The model file is not
synthesizable but due to the dependency, it is included in the syn
filelist. The synthesis fails at the analyze stage.

This commit is to fix the error by wrapping otbn_core_model.sv with
`OTBN_BUILD_MODEL`.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>