[otbn] Teach ISS to generate errors on call stack under/overflow
This is a reasonably large patch because the ISS didn't do any
handling of architectural errors before. We now define a special
exception type (Alert) in the Python and raise it if something goes
wrong. If that happens, it gets caught in OTBNSim.step(), which tells
the state to throw away any changes that were in flight and then to
come to a screeching halt.

The commit also comes with two simple ISS tests, to check that we stop
properly on underflow or overflow. Since the RTL doesn't support this
yet, we'll get mismatches for code that triggers the error (but
hopefully we don't actually have any such code in our testing at the
moment).

Finally, we add the new enum item to the DIF.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>