[otbn] Improve documentation of call stack and GPRs in general
I've got rid of the table, because I think it makes things harder to
understand. The call stack behaviour is now described carefully in the
GPRs section (before, it was rather duplicated), with links to the
"how to call and return" information in JAL/JALR.

This commit removes the note that GPRs and WDRs are separate register
files. I think this is pretty clear from the rest of the document, and
there's already a note in the previous paragraph to show which
instructions operate on which registers.

Finally, this commit registers a new named error code for call stack
errors in the otbn.hjson and otbn_pkg.sv and refers to the name in the
documentation.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>