[dv, meson] Integrate DV SW build steps into meson
This change is part 3/4 commit series that updates the way SW tests are
build for chip level. For the chip level SW tests to pass, all 4 commits
are needed.

This integrates the step that extracts the logs and rodata sections from
the embedded SW elf into meson.
This change is only enabled for 'sim_dv' device, so others should remain
unaffected.
Previously, this was done in DV simulation flow
`hw/dv/tools/dvsim/sim.mk`.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>