Knock OTBN memory range from 4MiB to 256KiB
This should make it a bit easier to fit all the peripherals together
in the memory map. There's still a 64KiB address space for each
memory (they're currently 4KiB in size), so we can grow by a factor of
16 without running out of addresses. If we're slapping down more SRAM
than that, our address map will be the least of our worries!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>