[OTBN DV] Prototype impl of gen-binaries
This builds on top of the previous commit to generate some binaries for
a test and a pre-run step. This is suitable in terms of reproducability
and wider coverage since the test seed is also passed on to the RIG
binary generator script. The downside of this approach is the generator
is invoked for each and every seed of each and every test that needs it,
which could be an overkill. It however does not seem to be a strain o
the simulation time though.

If a post build step is desired instead, then that can also be achieved
by moving the commands to `post_build_cmds` variable placed bare in the
HJSon to ensure it is executed for all available builds. The paths of
course, will need to be updated. The downside of this is the `--seed`
switch of the script would have to be foregoed.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>