[rom_ext] Remove Signature Algorithm Identifier
This field was being used to encode the Hash algorithm we had chosen. We
are using the PKCS1 v1.5 padding scheme, so this information is now
redundant.

I have chosen to leave a reserved field in the manifest here, instead of
moving all the other fields, as a prudent measure in case we need a
small field in the header in the future.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>