[dv, sw_test_status_if] Add sw_test_passed signal
This adds a `sw_test_passed` signal to make it easy for upstream
components to look it up to determine the final test pass / fail status
(for which the SW test status interface is one of the voters).

It also cleans up the intermediate SW test status values printed in hex
- earlier versions of Verilator did not support printing enum names in
string, which seems to have been fixed with 4.104.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>