[Verilator] Test pass/fail signature & banner
This change unifies the way the test status is determined and printed in
the simulation log file, across DV and Verilator. The DV report test
status method is invoked in top_earlgrey_verilator.sv after the pass /
fail signature is detected from SW. It prints a specific signature and
banner. The signature is used by DVSim to guage the test status.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>