[otbn] Remove dependency on riscv-python-sim
We were just using this for the Trace and TracePC classes, which
contain almost nothing. Define our own Trace/TracePC classes, remove
the "# type: ignore" hacks and fix the typing bugs that they hid.

Finally define an explicit Trace.trace function, rather than relying
on __str__ to generate trace output. We'll need this soon when we
start comparing with the RTL (because we'll want a different format
for different output types).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>