[kmac] Connect only one bit per 8bit in packer mask
The previous fix 003108511 removed the endian conversion logic and
directly connects the packer outputs to fifo_wdata struct.

The strobe in the fifo struct is byte strobe but the revised logic only
directly assigned the 64bit packer mask into 8bit byte strobe.

This commit is to connect only LSB in a byte to fifo strobe signal.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>