[otp_ctrl] Statically precompute hashed tokens from netlist constants
The all-zero token and the RAW unlock token are both netlist constants
for which we have to precompute the hashed token values such that they
can be used for the token comparisons inside the life cycle controller.

We rely on constant propagation to statically precompute these values
using the prim_prince cipher primitive. This precomputation is done
within the OTP controller since only here we have full access to the
random netlist constants used for the digest function.

Signed-off-by: Michael Schaffner <msf@opentitan.org>