[otbn] Correct ordering in flag enum
There's been a bit of confusion over the last couple of months about
the order (CMLZ or CLMZ). We came down on the former, but this enum
was still backwards.

Also, reorder the list of flags in the documentation to match. (It
wasn't technically wrong, but it's probably less confusing if it's the
same as everything else).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>