[aes] Prepare distribution of PRD for domain-oriented masking (DOM)
Unlike for the masked Canright S-Boxes, the output masks for SubBytes are
not known beforehand when using DOM. Instead, the output masks are
generated from the PRD input inside the individual S-Boxes.

In order to simplify the integration of different masking schemes, this
commit thus aligns how the masking PRD is distributed when using the masked
Canright S-Boxes. We no longer separate the masking PRNG output into output
masks and PRD for SubBytes at the cipher core level but feed the full
per-S-Box PRNG output into each S-Box. The "generation" of the output mask
is performed inside the S-Box and the output mask becomes an output port
also for the masked Canright S-Boxes.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>