[dvsim] Fix regression publish result link with --remote switch
If we use `--remote` switch, the result summary path is not correct.
This is because to remove the root directory and get `rel_path` in
FlowCfg, we used
```
        if self.rel_path == "":
            self.rel_path = os.path.dirname(self.flow_cfg_file).replace(
                self.proj_root + '/', '')
```
For this one, if remote switch is on, the self.proj_root is remote
repo_top. However, `self.flow_cfg_file` still uses the local abs path,
so the rel_path ended up still include the absolute path.
This PR fixed it by using the remote `cfg_path` if --remote is turned
on.

Signed-off-by: Cindy Chen <chencindy@google.com>