[kmac] Revise the endianness in KMAC
Previous commit 00310851134f35124264a55458a9fc9b1e48d645 fixed the
ordering issue. Based on the update, the KMAC top is revised w.r.t the
reverted endianness in MSG_FIFO.

The CFG.msg_endianness represents little-endian if the value is 0.
The KMAC top, however, byte-swapped the TL input data and mask if 0.

Also, the msg_endianness describes the value to affect secret keys. So,
the design is updated accordingly.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>