[prim_sync_reqack] Add variant with associated data and optional data reg
This commit adds a new variant of the REQ/ACK synchronizer primitive with
associated data. The data flow can be configured to be from source to
destination domain (push) or from destination to source domain (pull). In
the latter case, an optional register stage can be enabled to keep the
data stable until the ACK becomes visible in the source domain.

This is related to lowRISC/OpenTitan#4465.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>